+++ GameOffsets/ServerStashTabOffsets.cs
@@ -8,8 +8,8 @@ namespace GameOffsets
     {
         [FieldOffset(0x8)] public NativeStringU Name;
         [FieldOffset(0x2C)] public uint Color;
-        [FieldOffset(0x31)] public uint MemberFlags;
-        [FieldOffset(0x32)] public uint OfficerFlags;
+        [FieldOffset(0x31)] public byte MemberFlags;
+        [FieldOffset(0x32)] public byte OfficerFlags;
         [FieldOffset(0x34)] public uint TabType;
         [FieldOffset(0x38)] public ushort DisplayIndex;
         [FieldOffset(0x3D)] public byte Flags;
